Move pages around and refactor to better support rendering content outside the app body
